refactor: move shared types to own file
