+++ services/offline/src/lib/offline-provider.tsx
@@ -1,36 +1,8 @@
 import PropTypes from 'prop-types'
 import React from 'react'
+import { OfflineInterface } from '../types'
 import { CacheableSectionProvider } from './cacheable-section-state'
 import { OfflineInterfaceProvider } from './offline-interface'
-
-interface PromptUpdate {
-    (params: { message: string; action: string; onConfirm: () => void }): void
-}
-
-interface StartRecording {
-    (params: {
-        sectionId: string
-        recordingTimeoutDelay: number
-        onStarted: () => void
-        onCompleted: () => void
-        onError: (err: Error) => void
-    }): Promise<undefined>
-}
-
-interface IndexedDBCachedSection {
-    sectionId: string
-    lastUpdated: Date
-    // this may change in the future:
-    requests: number
-}
-
-interface OfflineInterface {
-    readonly pwaEnabled: boolean
-    init: (params: { promptUpdate: PromptUpdate }) => () => void
-    startRecording: StartRecording
-    getCachedSections: () => Promise<IndexedDBCachedSection[]>
-    removeCachedSection: (id: string) => Promise<boolean>
-}
 
 interface OfflineProviderInput {
     offlineInterface?: OfflineInterface
